Changed Path encoding scheme to emit strings instead or arrays, using escape sequences for invalid bytes.
In the common case of Paths containing valid unicode, this can be
a huge space optimization for certain encoders.

For example, with the json encoder the compact output size will
be reduced to something between 25% and 30% of the previous scheme:
Compare [47,104,111,109,101] and "/home".

If a Path consists _mostly_ out of non-ascii bytes and
either contains the escape symbol '%' or is not valid unicode,
then the output might end up bigger than before.
However, this seems like a rare case.