Check for zero len input when computing num_lines
If the user specifies that he wants to wrap the output at a particular
line length, then the memory preallocated is adjusted for the number
of newlines that are inserted. However, when the input has zero length,
then the num_lines calculation underflowed. So, we need to check for a
zero length input before calculating num_lines. If it is zero, then the
number of lines necessarily must be zero.