decrease rounding error when parsing f64 from JSON
As the precision of a floating point number depends
on its magnitude, adding each fractional digit to a large
number accumulates the floating point error quite quickly.

The error is accumulated much slower when assembling
the fractional part separately and then finally adding it
to the large number. Vide: the added test cases.